2026/day-44/requirements.txt
Flask==3.0.2
psycopg2-binary==2.9.9
psutil==5.9.8
